add src dir
