+++ .babelrc
@@ -4,5 +4,6 @@
   ],
   "plugins": [
     "@babel/plugin-proposal-class-properties"
-  ]
+  ],
+  "compact" : false
 }
