--- models/users.js
@@ -1,12 +1,0 @@
-// 'use strict';
-// module.exports = (sequelize, DataTypes) => {
-//   const Users = sequelize.define('Users', {
-//     firstName: DataTypes.STRING,
-//     lastName: DataTypes.STRING,
-//     email: DataTypes.STRING
-//   }, {});
-//   Users.associate = function(models) {
-//     // associations can be defined here
-//   };
-//   return Users;
-// };
