+++ package.json
@@ -2,11 +2,14 @@
   "name": "ekalaamu-backend",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "app.js",
   "scripts": {
-    "start": "NODE_ENV=production npm run migrate  && node --require babel app.js",
-    "dev": "NODE_ENV=dev nodemon --exec babel-node app.js",
-    "migrate": "NODE_ENV=production node_modules/.bin/sequelize db:migrate",
+    "start": "node ./build/app.js",
+    "dev": "NODE_ENV=dev nodemon --exec babel-node ./src/app.js",
+    "migrate": "node_modules/.bin/sequelize db:migrate",
+    "clean": "rm -rf build && mkdir build",
+    "build-babel": "babel -d ./build ./src -s && babel -d ./build/config ./config -s",
+    "build": "npm run clean && npm run build-babel",
     "test": "NODE_ENV=test",
     "sequelize": "dotenv sequelize"
   },
@@ -26,6 +29,7 @@
     "express": "^4.17.1",
     "express-promise-router": "^3.0.3",
     "install": "^0.13.0",
+    "node":"10.15.0",
     "jsonwebtoken": "^8.5.1",
     "npm": "^6.11.3",
     "pg": "^7.11.0",
